5449 - extract getData
